--- js/functions.js
@@ -1,64 +1,0 @@
-// Функция для проверки длины строки
-const checkLength = (string, length) => string.length <= length;
-checkLength('проверяемая строка', 20);
-checkLength('проверяемая строка', 18);
-checkLength('проверяемая строка', 10);
-
-
-// Функция для проверки, является ли строка палиндромом
-const isPalindrome = (string) => {
-  const tempString = string
-    .toLowerCase()
-    .replaceAll(' ', '');
-  let reverseString = '';
-  for (let i = tempString.length - 1; i >= 0; i--) {
-    reverseString += tempString.at(i);
-  }
-  return tempString === reverseString;
-};
-
-isPalindrome('топот');
-isPalindrome('ДовОд');
-isPalindrome('Кекс');
-isPalindrome('Лёша на полке клопа нашёл ');
-
-// Функция, которая принимает строку, извлекает содержащиеся в ней цифры от 0 до 9
-// и возвращает их в виде целого положительного числа
-
-const extractNumber = (string) => {
-  if (typeof string === 'number') {
-    return string;
-  }
-  let result = '';
-  for (let i = 0; i < string.length; i++) {
-    if (!Number.isNaN(parseInt(string.at(i), 10))) {
-      result += string.at(i);
-    }
-  }
-  return parseInt(result, 10);
-};
-
-extractNumber('2023 год');
-extractNumber('ECMAScript 2022');
-extractNumber('1 кефир, 0.5 батона');
-extractNumber('агент 007');
-extractNumber('а я томат');
-
-extractNumber(2023);
-
-
-// Функция, которая принимает три параметра: исходную строку, минимальную длину и строку
-// с добавочными символами — и возвращает исходную строку, дополненную указанными символами до заданной длины.
-
-const myPadStart = (string, minLength, pad) => {
-  const actualPad = minLength - string.length;
-  return actualPad <= 0
-    ? string
-    : pad.slice(0, actualPad % pad.length) + pad.repeat(actualPad / pad.length) + string;
-};
-
-myPadStart('1', 2, '0');
-myPadStart('1', 4, '0');
-myPadStart('q', 4, 'werty');
-myPadStart('q', 4, 'we');
-myPadStart('qwerty', 4, '0');
