+++ js/main.js
@@ -1,54 +1,3 @@
-const DESCRIPTIONS = [
-  'Просто фото',
-  'Красивое фото',
-  'Случайно получилось',
-  'Всем привет',
-  'Это я на отдыхе',
-  'Это я уже дома'
-];
-const MESSAGES = [
-  'Всё отлично!',
-  'В целом всё неплохо. Но не всё.',
-  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
-  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
-  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
-  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
-];
-const NAMES = [
-  'Артём',
-  'Иван',
-  'Кирилл',
-  'Мария',
-  'Анастасия',
-  'Кристина'
-];
-const OBJECTSQUANTITY = 25;
-const COMMENTSQUANTITY = 6;
+import {createPhotos} from './data.js';
 
-const getRandomInteger = (a, b) => {
-  const lower = Math.ceil(Math.min(a, b));
-  const upper = Math.floor(Math.max(a, b));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-
-const comments = () => ({
-  id: getRandomInteger(1, 500),
-  avatar: `img/avatar-${getRandomInteger(1, 6)}.svg`,
-  message: getRandomArrayElement(MESSAGES),
-  name: getRandomArrayElement(NAMES),
-});
-
-const photos = () => ({
-  id: getRandomInteger(1, 25),
-  url: `photos/${getRandomInteger(1, 25)}.jpg`,
-  description: getRandomArrayElement(DESCRIPTIONS),
-  likes: getRandomInteger(15, 200),
-  comments: Array.from({length: COMMENTSQUANTITY}, comments),
-});
-
-const createPhoto = () => Array.from({length: OBJECTSQUANTITY}, photos);
-
-createPhoto();
+createPhotos();
